+++ src/domain/event/formSections/registrationSection/RegistrationSection.tsx
@@ -17,7 +17,7 @@ import FieldColumn from '../../../app/layout/fieldColumn/FieldColumn';
 import FieldRow from '../../../app/layout/fieldRow/FieldRow';
 import Section from '../../../app/layout/section/Section';
 import useOrganizationAncestors from '../../../organization/hooks/useOrganizationAncestors';
-import { checkCanUserDoRegistrationAction } from '../../../registration/utils';
+import { checkCanUserDoRegistrationAction } from '../../../registration/permissions';
 import { REGISTRATION_ACTIONS } from '../../../registrations/constants';
 import useUser from '../../../user/hooks/useUser';
 import { EVENT_FIELDS } from '../../constants';
