+++ src/domain/registration/registrationForm/RegistrationForm.tsx
@@ -48,16 +48,14 @@ import ConfirmDeleteRegistrationModal from '../modals/confirmDeleteRegistrationM
 import RegistrationInfo from '../registrationInfo/RegistrationInfo';
 import styles from '../registrationPage.module.scss';
 import { RegistrationFormFields } from '../types';
-import {
-  checkCanUserDoRegistrationAction,
-  getRegistrationInitialValues,
-} from '../utils';
+import { getRegistrationInitialValues } from '../utils';
 import { getFocusableFieldId, registrationSchema } from '../validation';
 import getValue from '../../../utils/getValue';
 import GroupSizeSection from '../formSections/groupSizeSection/GroupSizeSection';
 import LanguagesSection from '../formSections/languageSection/LanguageSection';
 import RegistrationUserAccessesSection from '../formSections/registrationUserAccessesSection/RegistrationUserAccessesSection';
 import { useNotificationsContext } from '../../app/notificationsContext/hooks/useNotificationsContext';
+import { checkCanUserDoRegistrationAction } from '../permissions';
 
 export type CreateRegistrationFormProps = {
   event?: null;
@@ -100,7 +98,7 @@ const RegistrationForm: React.FC<RegistrationFormProps> = ({
   const isEditingAllowed = checkCanUserDoRegistrationAction({
     action,
     organizationAncestors,
-    publisher,
+    registration: registration ?? undefined,
     user,
   });
