+++ src/domain/registration/utils.ts
@@ -8,18 +8,15 @@ import isNil from 'lodash/isNil';
 import isNumber from 'lodash/isNumber';
 import omit from 'lodash/omit';
 
-import { MenuItemOptionProps } from '../../common/components/menuDropdown/types';
 import { NotificationProps } from '../../common/components/notification/Notification';
 import { FORM_NAMES, ROUTES, TIME_FORMAT_DATA } from '../../constants';
 import {
   CreateRegistrationMutationInput,
-  OrganizationFieldsFragment,
   RegistrationFieldsFragment,
   RegistrationQueryVariables,
   UpdateRegistrationMutationInput,
-  UserFieldsFragment,
 } from '../../generated/graphql';
-import { Editability, Language, PathBuilderProps } from '../../types';
+import { Language, PathBuilderProps } from '../../types';
 import { filterUnselectedLanguages } from '../../utils/filterUnselectedLanguages';
 import formatDate from '../../utils/formatDate';
 import formatDateAndTimeForApi from '../../utils/formatDateAndTimeForApi';
@@ -35,17 +32,6 @@ import i18n from '../app/i18n/i18nInit';
 import { getApiTokenFromStorage } from '../auth/utils';
 import { getEventFields } from '../event/utils';
 import {
-  hasAdminOrganization,
-  hasRegistrationAdminOrganization,
-  isAdminUserInOrganization,
-  isRegistrationAdminUserInOrganization,
-} from '../organization/utils';
-import {
-  REGISTRATION_ACTIONS,
-  REGISTRATION_ICONS,
-  REGISTRATION_LABEL_KEYS,
-} from '../registrations/constants';
-import {
   getSeatsReservationData,
   isSeatsReservationExpired,
 } from '../seatsReservation/utils';
@@ -58,153 +44,6 @@ import {
 
 export const clearRegistrationFormData = (): void => {
   sessionStorage.removeItem(FORM_NAMES.REGISTRATION_FORM);
-};
-
-export const checkCanUserDoRegistrationAction = ({
-  action,
-  organizationAncestors,
-  publisher,
-  user,
-}: {
-  action: REGISTRATION_ACTIONS;
-  organizationAncestors: OrganizationFieldsFragment[];
-  publisher: string;
-  user?: UserFieldsFragment;
-}): boolean => {
-  const isAdminUser = isAdminUserInOrganization({
-    id: publisher,
-    organizationAncestors,
-    user,
-  });
-  const isRegistrationAdminUser = isRegistrationAdminUserInOrganization({
-    id: publisher,
-    organizationAncestors,
-    user,
-  });
-
-  const hasAdminOrg =
-    hasAdminOrganization(user) || hasRegistrationAdminOrganization(user);
-
-  switch (action) {
-    case REGISTRATION_ACTIONS.CREATE:
-      return publisher ? isAdminUser || isRegistrationAdminUser : hasAdminOrg;
-    case REGISTRATION_ACTIONS.COPY:
-    case REGISTRATION_ACTIONS.COPY_LINK:
-    case REGISTRATION_ACTIONS.EDIT:
-    case REGISTRATION_ACTIONS.DELETE:
-    case REGISTRATION_ACTIONS.UPDATE:
-      return isAdminUser || isRegistrationAdminUser;
-    case REGISTRATION_ACTIONS.SHOW_SIGNUPS:
-    case REGISTRATION_ACTIONS.EDIT_ATTENDANCE_LIST:
-    case REGISTRATION_ACTIONS.EXPORT_SIGNUPS_AS_EXCEL:
-      return isRegistrationAdminUser;
-  }
-};
-
-export const getRegistrationActionWarning = ({
-  action,
-  authenticated,
-  registration,
-  t,
-  userCanDoAction,
-}: {
-  action: REGISTRATION_ACTIONS;
-  authenticated: boolean;
-  registration?: RegistrationFieldsFragment;
-  t: TFunction;
-  userCanDoAction: boolean;
-}): string => {
-  if (!authenticated) {
-    return t('authentication.noRightsUpdateRegistration');
-  }
-
-  if (!userCanDoAction) {
-    if (action === REGISTRATION_ACTIONS.CREATE) {
-      return t('registration.form.editButtonPanel.warningNoRightsToCreate');
-    }
-    return t('registration.form.editButtonPanel.warningNoRightsToEdit');
-  }
-
-  if (
-    registration &&
-    hasSignups(registration) &&
-    action === REGISTRATION_ACTIONS.DELETE
-  )
-    return t('registration.form.editButtonPanel.warningHasSignups');
-
-  return '';
-};
-
-export const checkIsRegistrationActionAllowed = ({
-  action,
-  authenticated,
-  organizationAncestors,
-  registration,
-  publisher,
-  t,
-  user,
-}: {
-  action: REGISTRATION_ACTIONS;
-  authenticated: boolean;
-  organizationAncestors: OrganizationFieldsFragment[];
-  publisher: string;
-  registration?: RegistrationFieldsFragment;
-  t: TFunction;
-  user?: UserFieldsFragment;
-}): Editability => {
-  const userCanDoAction = checkCanUserDoRegistrationAction({
-    action,
-    organizationAncestors,
-    publisher,
-    user,
-  });
-
-  const warning = getRegistrationActionWarning({
-    action,
-    authenticated,
-    registration,
-    t,
-    userCanDoAction,
-  });
-
-  return { editable: !warning, warning };
-};
-
-export const getRegistrationActionButtonProps = ({
-  action,
-  authenticated,
-  onClick,
-  organizationAncestors,
-  registration,
-  t,
-  user,
-}: {
-  action: REGISTRATION_ACTIONS;
-  authenticated: boolean;
-  onClick: () => void;
-  organizationAncestors: OrganizationFieldsFragment[];
-  registration?: RegistrationFieldsFragment;
-  t: TFunction;
-  user?: UserFieldsFragment;
-}): MenuItemOptionProps => {
-  const publisher = getValue(registration?.publisher, '');
-  const { editable, warning } = checkIsRegistrationActionAllowed({
-    action,
-    authenticated,
-    organizationAncestors,
-    publisher,
-    registration,
-    t,
-    user,
-  });
-
-  return {
-    disabled: !editable,
-    icon: REGISTRATION_ICONS[action],
-    label: t(REGISTRATION_LABEL_KEYS[action]),
-    onClick,
-    title: warning,
-  };
 };
 
 export const getRegistrationFields = (
