+++ src/domain/signups/signupActionsDropdown/SignupActionsDropdown.tsx
@@ -82,7 +82,7 @@ const SignupActionsDropdown: React.FC<SignupActionsDropdownProps> = ({
       authenticated,
       onClick,
       organizationAncestors,
-      publisher,
+      registration,
       t,
       user,
     });
